coord: added stat reporting, minor fixes

+++ include/coord/list.h
@@ -2,7 +2,7 @@
 #define LIST_H
 
 typedef struct Node {
-  int key;  // TODO: Check if causes problems
+  long key;  // TODO: Maybe switch back to int
   void* data;
   struct Node* next;
 } Node;
@@ -15,9 +15,9 @@ typedef struct {
 
 void ll_init(LinkedList* l);
 void ll_free(LinkedList* l);
-int ll_push_back(LinkedList* l, int key, void* data);
-int ll_push_front(LinkedList* l, int key, void* data);
+int ll_push_back(LinkedList* l, long key, void* data);
+int ll_push_front(LinkedList* l, long key, void* data);
 void* ll_pop_front(LinkedList* l);
-int ll_remove(LinkedList* l, int key);
+int ll_remove(LinkedList* l, long key);
 
 #endif
